break panel into new `GuiPanelTextures` + organize
